requirements.txt
paramiko==3.1.0
psutil==5.9.1
PyYAML==6.0
requests==2.22.0
